split concerns of lists
