--- src/components/Changesets_list.js
@@ -1,4 +1,0 @@
-import React from 'react';
-export function ChangetsetsList() {
-  return <p> hola</p>;
-}
